Make stream collecting work
Stream collect has Default + Extend trait requirements, which forces
the use of try_collect to collect results (which is a common use case of
iter collect), because Result does not satisfy these bounds.

The reason streams cannot use the same collection mechanism, IIUC, is
that async trait methods are not available (yet).